flask-python/proiect_db/python/src/auth/requirements.txt
autopep8==2.0.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
tomli==2.0.1
Werkzeug==2.2.2
